refactor: 상수 파일에서 레스토랑 데이터를 restaurantData.js로 분리
